# docs/openapi-identity.yaml
openapi: 3.1.0
info:
  title: Pulso Piura API
  version: 0.8.0
servers:
  - url: http://localhost:8080/api/v1
paths:
...
  /venues/{publicSlug}:
    parameters:
      - {in: path, name: publicSlug, required: true, schema: {type: string}}
    get:
      summary: Obtiene una sede publicada por su identificador público global
      responses:
        "200": {description: Sede publicada, content: {application/json: {schema: {$ref: "#/components/schemas/PublicVenue"}}}}
        "404": {$ref: "#/components/responses/NotFound"}
  /venues/{publicSlug}/spaces:
    parameters:
      - {in: path, name: publicSlug, required: true, schema: {type: string}}
    get:
      summary: Lista las canchas publicadas de una sede publicada
      responses:
        "200": {description: Canchas públicas, content: {application/json: {schema: {type: array, items: {$ref: "#/components/schemas/PublicSportSpace"}}}}}
        "404": {$ref: "#/components/responses/NotFound"}
...
  /organizations/{organizationId}/venues:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
    get:
      summary: Lista las sedes internas de una organización autorizada
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Sedes del tenant
          content:
            application/json:
              schema: {type: array, items: {$ref: "#/components/schemas/Venue"}}
        "403": {$ref: "#/components/responses/Forbidden"}
    post:
      summary: Crea una sede en estado borrador
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/VenueInput"}
      responses:
        "201":
          description: Sede creada
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Venue"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "409": {$ref: "#/components/responses/Conflict"}
...
  /organizations/{organizationId}/venues/{venueId}:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/VenueId"
    put:
      summary: Edita una sede mediante control optimista de versión
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/VenueUpdateInput"}
      responses:
        "200": {description: Sede actualizada, content: {application/json: {schema: {$ref: "#/components/schemas/Venue"}}}}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
    delete:
      summary: Archiva una sede sin eliminar su historial
      security: [{ bearerAuth: [] }]
      responses:
        "204": {description: Sede archivada}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
  /organizations/{organizationId}/venues/{venueId}/publish:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/VenueId"
    post:
      summary: Publica una sede borrador
      security: [{ bearerAuth: [] }]
      responses:
        "200": {description: Sede publicada, content: {application/json: {schema: {$ref: "#/components/schemas/Venue"}}}}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
  /organizations/{organizationId}/venues/{venueId}/spaces:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/VenueId"
    get:
      summary: Lista las canchas internas de una sede autorizada
      security: [{ bearerAuth: [] }]
      responses:
        "200": {description: Canchas de la sede, content: {application/json: {schema: {type: array, items: {$ref: "#/components/schemas/SportSpace"}}}}}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
    post:
      summary: Crea una cancha en estado borrador
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/SportSpaceInput"}
      responses:
        "201": {description: Cancha creada, content: {application/json: {schema: {$ref: "#/components/schemas/SportSpace"}}}}
        "400": {$ref: "#/components/responses/BadRequest"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    OrganizationId: {in: path, name: organizationId, required: true, schema: {type: string, format: uuid}}
    VenueId: {in: path, name: venueId, required: true, schema: {type: string, format: uuid}}
...
  responses:
    BadRequest: {description: Entrada inválida, content: {application/problem+json: {schema: {$ref: "#/components/schemas/Problem"}}}}
    Forbidden: {description: Sin membresía o permiso en la organización, content: {application/problem+json: {schema: {$ref: "#/components/schemas/Problem"}}}}
    NotFound: {description: Recurso no encontrado dentro del tenant, content: {application/problem+json: {schema: {$ref: "#/components/schemas/Problem"}}}}
    Conflict: {description: Estado, versión o dato en conflicto, content: {application/problem+json: {schema: {$ref: "#/components/schemas/Problem"}}}}
  schemas:
    Problem:
      type: object
      properties:
        type: {type: string, format: uri}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
    VenueInput:
      type: object
      required: [name, address, districtCode]
      properties:
        name: {type: string, maxLength: 160}
        address: {type: string, maxLength: 240}
        districtCode: {type: string, maxLength: 60}
        latitude: {type: number, minimum: -90, maximum: 90}
        longitude: {type: number, minimum: -180, maximum: 180}
        publicPhone: {type: string, maxLength: 30}
        amenityCodes:
          type: array
          uniqueItems: true
          items: {type: string, maxLength: 40}
    VenueUpdateInput:
      allOf:
        - $ref: "#/components/schemas/VenueInput"
        - type: object
          required: [version]
          properties:
            version: {type: integer, format: int64, minimum: 0}
    Venue:
      allOf:
        - $ref: "#/components/schemas/VenueUpdateInput"
        - type: object
          required: [id, organizationId, slug, publicSlug, status]
          properties:
            id: {type: string, format: uuid}
            organizationId: {type: string, format: uuid}
            slug: {type: string}
            publicSlug: {type: string, description: Identificador global estable para rutas públicas}
            status: {type: string, enum: [DRAFT, PUBLISHED, ARCHIVED]}
    SportSpaceInput:
      type: object
      required: [name, sportCode, formatCode, capacity, indoor]
      properties:
        name: {type: string, maxLength: 160}
        sportCode: {type: string, enum: [FOOTBALL, VOLLEYBALL, BASKETBALL, PADEL, TENNIS]}
        formatCode: {type: string, maxLength: 40}
        capacity: {type: integer, minimum: 1}
        surfaceType: {type: string, maxLength: 40}
        indoor: {type: boolean}
        amenityCodes:
          type: array
          uniqueItems: true
          items: {type: string, maxLength: 40}
    SportSpaceUpdateInput:
      allOf:
        - $ref: "#/components/schemas/SportSpaceInput"
        - type: object
          required: [version]
          properties:
            version: {type: integer, format: int64, minimum: 0}
    SportSpace:
      allOf:
        - $ref: "#/components/schemas/SportSpaceUpdateInput"
        - type: object
          required: [id, organizationId, venueId, status]
          properties:
            id: {type: string, format: uuid}
            organizationId: {type: string, format: uuid}
            venueId: {type: string, format: uuid}
            status: {type: string, enum: [DRAFT, PUBLISHED, ARCHIVED]}
...
    PublicVenue:
      type: object
      required: [publicSlug, name, address, districtCode, amenityCodes]
      properties:
        publicSlug: {type: string}
        name: {type: string}
        address: {type: string}
        districtCode: {type: string}
        latitude: {type: [number, "null"]}
        longitude: {type: [number, "null"]}
        publicPhone: {type: [string, "null"]}
        amenityCodes: {type: array, uniqueItems: true, items: {type: string}}
    PublicSportSpace:
      type: object
      required: [id, name, sportCode, formatCode, capacity, indoor, amenityCodes]
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        sportCode: {type: string}
        formatCode: {type: string}
        capacity: {type: integer, minimum: 1}
        surfaceType: {type: [string, "null"]}
        indoor: {type: boolean}
        amenityCodes: {type: array, uniqueItems: true, items: {type: string}}
...